Random strings length() = 2
I did some debugging and saw that the recursive call to generate() was in the wrong place, I fixed that and it crashed. Then I copied the array declaration from the top generator to the bottom one because that was the cause of the crash. Now it runs and generates two characters. I think the randomization in the bottom generator is not working.